move post to folder
