+++ instagram/static/instagram/src/post/Post.css
